requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.4.0
    # via httpx
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.1.2
    # via paramiko
cachetools==5.3.3
    # via google-auth
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
cosl==0.0.34
    # via -r requirements.in
cryptography==43.0.1
    # via paramiko
google-auth==2.28.1
    # via kubernetes
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.2
    # via lightkube
hvac==2.1.0
    # via juju
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via -r requirements.in
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
juju==3.5.2.0
    # via -r requirements.in
kubernetes==29.0.0
    # via juju
lightkube==0.15.4
    # via cosl
lightkube-models==1.30.0.8
    # via lightkube
macaroonbakery==1.3.4
    # via juju
markupsafe==2.1.5
    # via jinja2
mypy-extensions==1.0.0
    # via typing-inspect
oauthlib==3.2.2
    # via
    #   kubernetes
    #   requests-oauthlib
ops==2.16.0
    # via
    #   -r requirements.in
    #   cosl
    #   ops-scenario
ops-scenario==7.0.4
    # via pytest-interface-tester
packaging==23.2
    # via
    #   juju
    #   pytest
paramiko==3.4.0
    # via juju
pluggy==1.5.0
    # via pytest
protobuf==4.25.3
    # via macaroonbakery
pyasn1==0.5.1
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==2.9.1
    # via
    #   -r requirements.in
    #   cosl
    #   pytest-interface-tester
pydantic-core==2.23.3
    # via pydantic
pymacaroons==0.13.0
    # via macaroonbakery
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pytest==8.3.3
    # via pytest-interface-tester
pytest-interface-tester==3.1.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via kubernetes
pytz==2024.1
    # via pyrfc3339
pyyaml==6.0.1
    # via
    #   cosl
    #   juju
    #   kubernetes
    #   lightkube
    #   ops
    #   ops-scenario
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tenacity==9.0.0
    # via cosl
toposort==1.10
    # via juju
typer==0.7.0
    # via pytest-interface-tester
typing-extensions==4.10.0
    # via
    #   cosl
    #   pydantic
    #   pydantic-core
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==2.2.2
    # via
    #   kubernetes
    #   requests
websocket-client==1.7.0
    # via
    #   kubernetes
    #   ops
websockets==12.0
    # via juju
